# Tweak performance of E-Ink Spectra 6