The E-Ink Spectra 6 is likely not the fastest EPD, however it can be tweaked in case on knows the commands and arguments. With a bit of research, I determined the following: Especially the PLL CONTROL setting makes a huge difference, while the AUTO command can speed up the PON->DRF->POF(->DSLP) with pure hardware.

To give an impression, what the PLL CONTROL setting can do:

- E-Ink Spectra 6 - 4.0'
```
  D7 D6 D5 D4 D3 D2 D1 D0
  -  -  -  - [    FRS    ] -> Frame Rate Setting (in Hz)

  0x00 -> 37662ms (= 0x10)
  0x01 -> 30038ms
  0x02 -> 22091ms
  0x03 -> 19083ms (= 0x08 ~ 0x0F)
  0x04 -> 14809ms
  0x05 -> 12887ms
  0x06 -> 11448ms
  0x07 ->  9819ms
```
The default was 0x08 :-(
**With the new setting the 4.0' is below 10 seconds!**

- E-Ink Spectra 6 - 7.3'
```
  D7 D6 D5 D4 D3 D2 D1 D0
  -  -  -  - [    FRS    ] -> Frame Rate Setting (in Hz)

  0x00 -> 25365ms (= 0x10)
  0x01 -> 19863ms
  0x02 -> 14944ms
  0x03 -> 12932ms (= 0x08 ~ 0x0F)
  0x04 -> 10075ms
  0x05 ->  8798ms
  0x06 ->  7832ms
  0x07 ->  6746ms
```
The default was 0x03 :-(
**With the new setting the 7.3' is achieving 6.7 seconds!!!**

- E-Ink Spectra 6 - 13.3'
```
  D7 D6 D5 D4 D3 D2 D1 D0
  -  -  -  - [    FRS    ] -> Frame Rate Setting (in Hz)

  0x00 -> 76357ms (= 0x10)
  0x01 -> 38650ms
  0x02 -> 19681ms (= 0x08 ~ 0x0F) // seems default (compared to timing)
  0x03 -> Picture just dark
  0x04 -> 12789ms, but not always correct
  0x05 -> Picture not finished
  0x06 -> -''-
  0x07 -> not working
```
The default was 0x02 :-)
I guess, in case of the 13.3', the Waveshare adapter solely supplies a 3.3V with ~330mA (if I am not mistaken). That might be insufficient to refresh faster.

In case someone is interested in a much cleaner implementation, that can reach a bit faster speeds: https://codeberg.org/psiegl/E-Ink_Spectra_E6/